+++ packages/odata-service-inquirer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/odata-service-inquirer
 
+## 0.5.48
+
+### Patch Changes
+
+-   Updated dependencies [eb38e5b]
+    -   @sap-ux/project-access@1.28.2
+    -   @sap-ux/axios-extension@1.16.6
+    -   @sap-ux/telemetry@0.5.36
+
 ## 0.5.47
 
 ### Patch Changes