+++ packages/ui5-library-writer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/ui5-library-writer
 
+## 0.5.26
+
+### Patch Changes
+
+-   Updated dependencies [eb38e5b]
+    -   @sap-ux/project-access@1.28.2
+    -   @sap-ux/ui5-application-writer@1.2.0
+
 ## 0.5.25
 
 ### Patch Changes